Fix commands not being handled by QMK because it's one-byte short
The hidapi crate expects the first byte to be the report ID, and then
passes through the rest of the buffer as the data/message. This means
that the message length received by QMK would be 31 bytes. I guess QMK
at one point of time was fine with this, but it now expects it to be 32
bytes exactly.

Fix this by adding another byte.

Signed-off-by: Saikrishna Arcot <saiarcot895@gmail.com>